Vlaser_VLA/RoboTwin/policy/internvla_2B_parallel_decoding/requirements.txt
accelerate==1.6.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asciitree==0.3.3
asttokens==3.0.0
blobfile==3.0.0
chardet==5.2.0
cloudpickle==3.1.1
colorlog==6.9.0
configargparse==1.7.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
deepspeed==0.16.7
diffusers==0.32.2
distro==1.9.0
einops==0.8.1
einx==0.3.0
exceptiongroup==1.3.0
executing==2.2.0
farama-notifications==0.0.4
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.58.4
frozendict==2.4.6
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
gitpython==3.1.44
gymnasium==0.29.1
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
imageio==2.34.2
imgaug==0.4.0
importlib-resources==6.5.2
iopath==0.1.10
ipython==8.37.0
jaxtyping==0.3.2
jedi==0.19.2
jiter==0.10.0
joblib==1.5.1
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-core==5.8.1
kiwisolver==1.4.8
lazy-loader==0.4
lightning==2.5.1.post0
lxml==5.4.0
mapbox-earcut==1.0.3
matplotlib==3.10.3
matplotlib-inline==0.1.7
mplib==0.2.1
mpmath==1.3.0
nbformat==5.10.4
networkx==3.4.2
ninja==1.11.1.4
numcodecs==0.13.1
numpy==1.26.4
numpy-quaternion==2024.0.8
omegaconf==2.3.0
openai==1.88.0
opencv-python==4.11.0.86
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
portalocker==3.2.0
prompt-toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pybind11==2.13.6
pycollada==0.8
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
rpds-py==0.25.1
rtree==1.4.0
safetensors==0.5.3
sapien==3.0.0b1
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.30.0
setproctitle==1.3.6
shapely==2.1.1
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
svg-path==6.3
sympy==1.13.1
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
toppra==0.6.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
transforms3d==0.4.2
trimesh==3.21.7
triton==3.2.0
typeguard==4.4.4
typing-extensions==4.14.0
typing-inspection==0.4.1
wadler-lindig==0.1.7
wandb==0.20.1
warp-lang==1.4.0
wcwidth==0.2.13
xxhash==3.5.0
yacs==0.1.8
yourdfpy==0.0.57
zarr==2.18.3